refact: Initial Migration and BaseEntity

+++ AMS.Application/UseCases/Users/Command/DeleteUser/DeleteUserHandler.cs
@@ -1,13 +1,16 @@
-﻿using AMS.Application.Commons.Bases;
+﻿using System.Security.Claims;
+using AMS.Application.Commons.Bases;
 using AMS.Application.Commons.Utils;
 using AMS.Application.Interfaces.Persistence;
 using MediatR;
+using Microsoft.AspNetCore.Http;
 
 namespace AMS.Application.UseCases.User.Command.DeleteUser
 {
-    public class DeleteUserHandler(IUnitOfWork unitOfWork) : IRequestHandler<DeleteUserCommand, BaseResponse<bool>>
+    public class DeleteUserHandler(IUnitOfWork unitOfWork, IHttpContextAccessor httpContext) : IRequestHandler<DeleteUserCommand, BaseResponse<bool>>
     {
         private readonly IUnitOfWork _unitOfWork = unitOfWork;
+        private readonly IHttpContextAccessor _httpContext = httpContext;
 
         public async Task<BaseResponse<bool>> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
@@ -15,6 +18,15 @@ namespace AMS.Application.UseCases.User.Command.DeleteUser
 
             try
             {
+                var userId = _httpContext.HttpContext.User.Claims.FirstOrDefault(x => x.Type == "IdEntidad")?.Value;
+
+                if (string.IsNullOrEmpty(userId))
+                {
+                    response.Status = (int)ResponseCode.UNAUTHORIZED;
+                    response.Message = ExceptionMessage.RESOURCE_NOT_FOUND;
+                    return response;
+                }
+                
                 await _unitOfWork.UserRepository.DeleteAsync(request.Id);
                 response.Status = (int)ResponseCode.OK;
                 response.Message = ResponseMessage.DELETE_USER_SUCCESS;
